HTML Tree now works
You can toggle the view of the HTML Tree in "View > HTML Tree"
This will bring up a draggable window that shows all the elements in the document in an interactive <svg>. Clicking one of the elements will select that element in the document (this allows you to select elements that would normally be covered up by other elements and are thus unselectable). Elements don't show their name so it might be kind of hard to tell which element is which but I'll try to fix that in another update.