position:fixed now supported!
Fixed elements will no longer incorrectly scroll with the document and
will remain in their spot, as you would expect from a fixed element. The
main <iframe>'s height was changed, its vertical scroll bar was removed,
and it is now manually scrolled depending on #framecontainer's scroll.
This works well with most computer browsers but it will be kind of jerky
on mobile devices where the scroll event is only fired after the user
has stopped scrolling. Then again, this page is meant for computer
browsers only pretty much so it should be fine.

A google analytics script was added.

A Modernizr script was also added to detect browser support for
position:sticky. This makes for smoother scrolling on some browsers.

The `Save Styles` button for `[Right Click] > Edit Styles...` now works.

The background image in presets/3.html was changed since fixed elements
are now supported in the document editor.