Completely remade `Edit Text`
A bunch of bugs have been fixed and a ton of other improvements.

HTML entity replacers are made in DocumentFragments now instead of the
actual document for performance enhancements.

When an HTML entity replacer is clicked, only the text node containing
the entity sequence is altered. This means, if the user's caret is
editting another text node, the caret won't be moved and the user can
continue editting.

The user can now see the caret when it is in between an HTML entity
sequence.

Instead of appearing as an ugly blue blob, HTML entity replacers change
the text color which looks cleaner while still grabbing the user's
attention.

The overlay isn't updated every keypress as it was previously; instead
it is only updated once the user is finished editing, allowing for way
faster performance.

The Open Sans googleapis.com css file is loaded once the page has
finished rendering to improve page load time.

Overall, editing the inner text of an element is more reliable, fast,
and user-friendly.