A bunch of stuff
Dialogs will now be shrunken down if their content isn't large enough to
fill up the default height. This prevents half-empty dialogs.

Added an option to permanently hide the dialog that appears when the
user is unable to paste.

Added a new dialog that displays the clipboard shortcuts. This only
appears when the user tries to click Copy, Cut, or Paste in a context
menu and it fails.

Added an exit button to the HTML Viewer so that the user doesn't have to
go to View > HTML Tree each time.

The Undo and Redo options in the Edit context menu show the action that
will be undone/redone. (e.g. it will sometimes show "Undo • Edit text"
or "Redo • Update document title") to give the user more information.

Added Paste option to the right click context menus.

A variable stores the previously copied value so that the user has a
sort-of backup option if their browser doesn't support pasting. This
does mean that the user isn't able to copy from an external source, but
it's still better than not being able to paste at all.