Fixed two bugs
On browsers with no support for position:sticky, an error would be thrown since I forgot to include "this." in front of scrollTop (line 2705)

When a first-time user starts the page for the first time, no history entries are defined. An error would be thrown then on certain lines of code where the script tries to get .action from `undefined`.