Fixed another bug
When the user tries to paste using the context menu and document.queryCommandSupported('paste') (incorrectly) returns true, sometimes document.execCommand('paste') will still fail. In this scenario, the user is not able to paste and the context menu showing the keyboard shortcuts does not appear. Now, if the command is not executed, the context menu will still appear, even if the 'paste' command is technically supported. This also applies to copy and cut.