Added Insert Child option
There was always just an Insert Child option in the right click context
menu, but it never did anything, even though I never disabled it. I
finally added some functionality to that option.
Clicking Insert Child will open a dialog where the user can select a
node name for the child being inserted. It can be an HTMLElement or a
Text Node. Adjacent text nodes are automatically combined. At the bottom
of the dialog, the user can select where the new node should be in
relation to its siblings. Text nodes and HTML elements are shown so that
the user can place the child before or after a text node if they want.
That's the big change. A bunch of other small stuff was changed as well.

CSS rules that aren't immediately needed by the browser to display
above-the-fold are now placed in other CSS files that are loaded once
the main content is able to be displayed. This is mostly for speed
improvements and organization.

A few SVGs were added to serve as context menu icons.

Fixed a small bug with Edit Text.

Fixed a bug where the main overlay element's alias attribute was not
being updated when the user undid or redid changes. The alias was always
set to the previous version and it would cause some problems. The alias
is set each time overlayUpdate is called now.

When the overlay is updated, the user's previous selection is maintained
now (most of the time; the user's selection can't be saved when
undoing/redoing because new element's are replacing the user's old
selection and the script doesn't have a reference to the new elements
yet)

Fixed a small bug with restoring the user's selection

Changed the way the title of the window is made. ([USER TITLE] - HTML
Studio · ChristianFigueroa.GitHub.io)

When the user undoes/redoes a change in the document title, the width of
the textbox is automatically changed to match the new value.