Better context menu keyboard control, new unwrap element option, and bug fixes
Fixed a typo from pesudoContextMenuItem to pseudoContextMenuItem

The sub menus (File, Edit, etc.) now have better key board control:
- The user can now tab to select a submenu button and press enter to
open/close it.
- The user can also use down arrow key to open the context menun and
select the first item
- All four arrow keys are used in context menus now:
- Down selects the next item down, or the first item if there are no
more items below the current one
- Up selects the next item up, or the last item if there are no more
items above the current one. With submenus, the text to open the submenu
is selected instead (Edit is selected instead Preferences inside the
Edit submenu)
- Left will select the subcontextmenu of the item if one is defined for
the item
- Right will select the subcontextmenu's parent item if one is defined
- The arrow kes also work for the text buttons that open submenus

- New Ctrl + H shortcut to edit the HTML of an element

- New Unwrap Element option in the right click context menu. The
selected node is replaced with all its children (useful for replacing
links or other inline elements with the text inside of the link)

The right click context menu is no longer ~0.3em too long on the right
side

- ContextMenu objects now have an open and close method that will make
the contextmenu appear or disappear respectively

- Made updateOverlay() slightly faster by appending the child nodes to
their parent BEFORE adding the parent to the DOM (prevents the browser
from updating the DOM as much)

- Adjacent text nodes are automatically combined for each call to
updateOverlay

- Removed a bug where adding an element to your current selection would
quickly make a selection range (which colors the text of the page blue
and is the default behavior of most browsers) and then delete it

- Most keyboard shortcuts are now disabled if a dialog is open and
blocking the document editor

- ContextMenu.js is now unminified because it's such a pain to edit it
in its minified form. It will eventually be replaced again with a
minfied version for faster loading time, but not while it's still being
worked on