Page load time improvements
The super big main script that controls pretty much everything is now
stored under script/main.js and is deferred on page load to improve the
page's load time. While the page is loading, things like the sub menu
buttons (File, Edit, etc.) are grayed out to show that you can't click
on them yet (not until the script has loaded).

Fixed a bug with the <noscript> containing all the deferred CSS <link>s.
It used to cover up the message the user would see if they had their
JavaScript disabled and it just looked like an empty page. It's z-index
is set to -1 now so that the actual <noscript> will appear above it.

New saveAs and Blob libraries that will be used for downloading the
user's document.

Almost all <script>s now are deferred so they don't block the borwser
from rendering the page.