More Edit Text Improvements
Replaces can now be auto-clicked by pressing tab while typing inside the
replacer. For example, type "&copy;" then press tab and the &copy; will
automatically be replaced with © so that the user can continue typing
without having to stop what they're doing just to click the text. This
also applies to element replacers (i.e. typing <a> to create an anchor
element).

When the user replaces text using a replacer (as talked about above),
their selection is now saved. The caret will no longer be taken to the
beginning of the text node (which used to force the user to go back to
their selection manually). The selection will also be saved when the
user has characters selected (i.e. the selection isn't collapsed).

On browsers supporting document.execCommand, the user can now undo/redo
changes made by replacers. Previously, the browser had no way of knowing
that the text was being changed programmatically, so pressing Ctrl + Z
wouldn't undo the replacement, and some of the text would be undone
while some wouldn't. Now, the user can reliably undo/redo even when they
are interacting with the replacers.

Users can now also use shortcuts Ctrl + B to bold text and Ctrl + I to
italicize text while editing the text of an element.

Fixed some bugs on chrome with keypress vs keydown events.

Fixed some other bugs with the semicolon character having different
keyCodes in different browsers.

Entity replacers (e.g. &amp; &#123;) are also recognized when they are
followed by white-space or at the end of the text node. This means the
user can type in "Some text &amp more text" and the "&amp" will be
recognizesd as an entity even though it is not followed by a semicolon.

Users can now double click elemenets as a shortcut to edit the text of
an element.

Fixed a bug with the width of the title <input> not expanding correctly
when empty.