Minor bug fixes and improvements
The <span> elements that show the dimensions of selected elements while
View > Bounding Boxes is enabled look a little better than the old plain
white boxes.

Fixed a bug with Ctrl + Shift + Left/Right Arrow. When the user tried to
select the next sibling of an element, if it was the last child of its
parent, the first child is supposed to be selected. Previously, it just
threw an error because it was selecting the wrong element. It correctly
selects the first child now (or last when you're going in reverse).

Added a filter for Unwrap Element. If the <body> is the selected
element, Unwrap Element appears disabled in the context menu since you
can't unwrap the <body>.

Fixed selecting elements so that it works faster when bounding box are
disabled. Previously, it took longer to select elements because their
bounding box was calculated even if View > Bounding Boxes was disabled.
Now, it should work faster if View > Bounding Boxes isn't enabled.