Fixed annoying CSS specificity issue
Basically, some CSS styles weren't being applied when they were supposed
to be because of their order in the stylesheet. This fixes that. Read on
for more technical talk.

Stylesheets are parsed by HTML Studio so that they can be used when
styling the overlay. Basically, it finds all the styles that pertain to
an element and applies those styles directly onto the element's style
attribute so that the element inherits the same styles as the element
its mimicking (in the <iframe>). This helps to keep the overlay in sync
with the actual document. This caused a problem though so since
different CSS selectors have different levels of specificity. This means
"div.class" takes precedence over just "div" or ".class". Previously,
specificity wasn't considered when styles were being applied to the
elements and it would occasionally cause them to be out of sync with the
overlay. This update fixes that bug so that elements are in sync with
their counterparts. The paragraph below is mostly code-specific changes.

CSS selectors are given a specificity amount based on the number of IDs,
classes, and element names used in the selector (and other things like
pseudo-classes, but they are parsed the same way).
script/parseSelector.js adds a function that, given a stringified CSS
selector, returns the specificity for that selector. Specificities are
now compared when HTML Studio is applying styles to elements from
stylesheets now so that the most specific selectors "beat" the
less-specific selectors. Certain selectors like "div, span" are also
split up now and given two separate specificity values since there are
technically two selectors in "div, span" (separated by commas).
Something like "div#id.class, span {color: black;}" is stored as
"div#id.class {color:black;}" and "span {color:black;}" since the two
selectors have different specificity values and thus have to be applied
differently.

Besides all the other stuff above, a bug was also fixed in Chrome where
users couldn't edit stylesheets because the dialog would open an empty
sheet instead of the one with all the styles.