Improvements to the stylesheet editor
A bunch of improvements were made to the dialog that open when you go to
Style Sheet > Create New.

Using the same API from the previous commit to calculate CSS
specificity, CSS selectors can now be syntactically highlighted. If the
user enters an invalid CSS selector for example, the selector will
appear red or will appear green for IDs, etc. Highlighting can also be
toggled on or off using a checkbox labeled "Syntax Highlighting."

A lot of improvements were made to the selector parser to make it more
accurate (e.g. recognizing "#\ " as an ID with a space instead of just
trimming the space off).

While editting the value of a CSS style, the user can now use multiple
lines. If the user types something, ends with a comma (to add another
value, like in box-shadow for example), and presses enter, a new line
will be added instead of just focusing the next element. Not a big
game-changer, but still nice to have to keep your CSS more organized.

Fixed some bugs in chrome with semicolons not being recognized with
keyCode 59.

Fixed a bug where the style sheet editor would sometimes open an empty
stylesheet instead of the actual stylesheet the user clicked on.

For browsers that support it, the "+" and "X" buttons that appear when
hovering over a CSS style will also appear when one of the <input>s
inside the style are focused.  (uses ":focus-within" CSS pseudo-class)

Context menu items with long names can now be wrapped using hyphens.

In the stylesheet editor, the value <input>s will now auto-espand to
fill all the available space. This helps the user when editing long
values by letting them see more of it.

Preparations for horizontally-resiazble dialogs.