Code Improvements and Color Selectors
Edit > Preferences > Selected Element Color (previously Node Selection
Color) has been implemented. The user can now choose which color
selected elements will apear as in case the default color conflicts with
the user's theme and is hard to see (or if the user just want to change
the color because they like it better).

Preferences can now be saved into localStorage so that they will persist
through multiple sessions instead of being reset just for reloading the
page.

New HTMLStudio.ColorSelector "class" that creates an interactive color
selector (which is what appears for the Selected Element Color option
mentioned above).

New wrapper for localStorage so that there's not just random calls to
localStorage.getItem or localStorage.setItem throughout the code. This
can also help in the future if I ever want to switch to a data base
instead of saving locally on the user's browser. This also helps because
the wrapper can pre-parse data instead of just returning the stringified
version of the data (i.e. no more unnecessary calls to JSON.parse and
JSON.stringifiy).

Even though preferences can now be saved between sessions, only the
selected element color preference has been implemented. All the other
ones are still reset when the page reloads. This will be implemented in
the next update.