<Table> updates and bug fixes
Context menus for <table> elements (e.g. <td>, <tr>, etc.) now have
extra options.
- For <td> and <tr> elements, there are Insert Row Above and Insert Row
Below options that will copy the selected row's cells (i.e. Merged cells
will be extended, single cells will be copied, spaces without a cell
won't be copied).
- For <td>, <col>, and <colgroup> elements, there are Insert Column Left
and Insert Column Right options that work exactly the same as Insert Row
Above/Below.
- For <table>, <thead>, <tbody>, and <tfoot> elements, there are Prepend
Row, Append Row, Prepend Column, and Append Column that do exactly as
their name suggests.
- All these options will also work for merged cells (colspan, rowspan),
intersecting cells, and areas of the table without a cell (no <td>
present).

<A> and <img> elements also have new context menu options: Edit href
Attribute and Edit src Attribute, respectively. This lets the user
quickly edit these attributes since they are so important to these types
of elements. Both these options are bound to the Ctrl + Shift + 8 (Ctrl
+ *) keyboard shortcut.

<a> elements also have an "Open in New Tab" option in their context menu
to allow the user to open the link in a new tab.

Added Ctrl + Q and Ctrl + E for Edit Styles and Edit Attributes options,
repectively.

Added Ctrl + Shift + I as shortcut to insert child.

Added Ctrl + U as shortcut to unwrap element.

Fixed a bug where the vertical scroll bar would ALWAYS appear in the
document editor, even if the document wan't big enough to be scrolled.

Fixed a bug in the stylesheet editor where typing :not( as a selector
would cause the syntax highlighting underlines to appear uneven (a small
red underline would appear with no innerText and throw everything else
off).

Added some some stuff to ContextMenu.js to allow for items in the
context menus to "disappear" when disabled.

Fixed bug in DraggableElements.js. I don't know what caused the problem,
but it seems to be fixed now.

Fixed a bug when editing the text of an element where the element's
borders would show up when they're not supposed to.

Fixed a bug when editing the attributes of an element where pressing "="
wouldn't focus the next element in Chrome.

Fixed a bug when inserting a child into an element where sometimes the
child would be inserted at the end of the element, even when the user
chose to place it at the beginning of the element.

Finally added Ctrl + S as a shortcut to download the file.

Added a Print option under the File submenu. Ctrl + P isn't working at
the moment since it tries to print the whole page instead of just the
user's document.

Ficed a bug where preferences weren't being saved correctly.

Fixed a bug where blacklisted elements' (self-closing elements and SVG
elements) text could be edited in certain situations when they're not
supposed to be.

When selecting elements from the HTML Tree, if there is a context menu
open, it will automatically update to match the new selection instead of
staying with the user's previous selection.

Pressing tab when editing HTML will now be a part of the brower's undo
stack (i.e. pressing Ctrl + Z will undo the tab instead of the user's
previous action).

Improvements to the way elements are highlighted when selected.

Fixed a bug where Shift clicking something from a context menu would
quickly select the entire page and then deselect everything. Nothing is
selected now when shifting.

A few other small improvements.