Improvements and stuff
The HTML Editor (Right Click > Edit as HTML) now has better editing.
- Tabbing when multiple lines are selected now increases the indentation
for each selected line instead of replacing the selection with a tab
character.
- When the user presses Enter to insert a new line, the new line
automatically inherits the same indentation as the line being Entered on
so that elements have the same indentation.
- Pressing Home will now move the cursor to the beginning of the text of
the line, not the beginning of the line itself. This is to allow for
faster editing of the text on the line instead of editing the
indentation on the line.

When the HTML Editor for an element is opened, the text is now
automatially prettified to make it more legible for the user. The
prettification takes the CSS white-space property into account so that
white space is respected when prettifying.

CSS properties are now auto-filled into textboxes in the stylesheet and
style editor. For example, typing "bo" will fill in the textbox "border"
to allow the user to simply tab to autocomplete. Most CSS properties are
stored in an array in script/CSS_properties.json. CSS colors are also
autocompleted in the Color Editor (Edit > Preferences > Selected Element
Color) (e.g. typing "re" will show an autocomplete for "rebeccapurple").

Typing CSS colors into the textbox mentioned above will now set the
opacity to 100% instead of keeping the user's previous opacity. This
feature seemed more annoying than helpful.

Fixed a bug in the Stylesheet Editor where typing Enter in an empty CSS
property value textbox would enter a new line character instead of
focusing the next elemenet.

script/minifyHTML.js was changed to script/formatHTML.js since it
includes both a minification function and a prettification function.
Both functions are stored under window.HTMLStudio.formatHTML.

Added a fail-safe for the four <table>-specific options Prepend/Append
Row/Column. If the table doesn't have any <tbody>s, a new row/colum is
automatically inserted into the <table> anyway.

After saving the HTML from the HTML Editor, all inserted top-level
elements are selected now instead of just deselecting all elements.

Fixed a bug where sometimes pressing Ctrl + Up or Ctrl + Down would do
nothing.