Small update
Added an icon for all major platforms.

A few more improvements to the HTML Editor.
- Line numbers are now displayed
- The selection is now scrolled into view on Home, Enter, and Tab as if
they were native to the browser

Lots of improvements to the HTML prettifier for less mistakes that
affect the white-space of elements.

Improvements to the gradient generator to change the colors of selected
elements.

In the color selector <input> in Edit > Preferences > Selected Element
Color, any separator is now allowed to separate numbers. Previously only
"0, 0, 0" or "0 0 0" would be valid. Now, "0; 0; 0", "0 $ 0 ) 0" will
also work since most non-number characters are recognized as separators.

Buttons in dialogs are now shown as being hovered on when the user
focuses the elements using Tab.